+++ test/scopedFixtures/stylesAttribute.test.js
@@ -103,6 +103,15 @@ describe('Scoped Mode', () => {
     expect(output).toBe(`<style module="scoped">.red-123 { color: red; }</style><span class="red-123" data-color="red">Red</span>`);
   });
 
+  test('Skip empty class attribute', async () => {
+    const output = await compiler({
+      source: `<style module="scoped">.red { color: red; }</style><span class="">Red</span>`
+    }, {
+      localIdentName: '[local]-123',
+    });
+    expect(output).toBe(`<style module="scoped">.red-123 { color: red; }</style><span class="">Red</span>`);
+  });
+
   test('Parse extra attributes as well', async () => {
     const output = await compiler({
       source: `<style module="scoped">.red { color: red; }</style><span class="red" data-color="red">Red</span>`